Remove externs and add them to gitignore
lein generate-externs need externs directory to be present

Create an externs directory with a gitignore file that only allows itself to
be commited

We are using simple optimizations for prod builds so this file is not used.

We have much more useful output with :simple optimizations if an error
happens, not just something like a.b.adsfas is undefined.

The difference in size between advanced and simple optimizations is only
about 300KB, and the RN packager also makes some magic with js after cljs
compiler anyway.