Add latch to ensure application messenger is set.
There are two ways that messages can reach the StatusService,
application events (which are handled via `handleEvent`) and
signal events (which are handled via `signalEvent`).

If a signal event fires before an application event, it still
uses handleEvent logic, and handleEvent will try to reply to
the message. If the `applicationMessenger` isn't set, the reply
fails and the react native app is never told that the status
node is running, which is why the app gets no messages.

I've added a CountDownLatch, forcing the event handler to wait
for the applicationMessenger to be set before replying to any
messages.

Tested by injecting a long wait into `handleMessage` and observing
the same failure we see in SauceLabs runs.