Fix and uncomment tests in contacts.events and utils.pre-receiver
I have uncommented the two namespaces:

contacts.events was mainly an issue with schema validation
and a few features not covered by tests.

the tests in pre-receiver were actually passing, the issue was a
test suite in wallet.events which was run asychronously by re-frame
which meant the http request in the fx was actually being fired.

Signed-off-by: Oskar Thoren <ot@oskarthoren.com>