[ISSUE #3706] collect mixpanel events while offline
This adds a tracking interceptor to `default-interceptors` that replaces the old
`add-post-event-callback` hook. The interceptor gets the required data from
app-db to know if it's online and passes that to a new mixpanel tracking fn that
queues up tracking events, and drains that queue iff the app is online again.

Signed-off-by: Julien Eluard <julien.eluard@gmail.com>