[#5985] Add animation for sending and receiving messages in a chat scene.
Because scaling in RN is anchored at the center([0.5, 0.5]) of a view, simultaneous translations along both the X and Y axis are required to achieve the desired effect of animating from the lower left/right point.

A new flag animate? has been added to each message to control whether it should show with animation.