Add contact request message
When pfs is enabled, if a contact code has not been already fetched, the
user will be presented with a different input, and it will be asked to
sent ad contact request message instead.

This message will be displayed differently and it will give the user an
option to accept the contact request, or block the contact.

Currently this PR does not limit the number of contact requests that can
be sent, it will be done in a later PR.
All the code is behind a the pfs toggle feature flag.