fix(signing/views): allow discarding signing popup when tx in progress
As discussed in #9977, there's an issue that the signing view popup cannot
be cancelled/discarded when a transaction is in progress.
This is because the `cancel-is-pressed` event is only conditionally applied to
the cancel button, rendering it non-functional for the time-being.

This commit changes that behaviour to always attach the event handler to
the cancel button, so that the popup can be closed even when a transaction
has been sent.

Fixes #9977