Added RefreshControl
To be used when pulling to refresh the wallet view.

Conditional loading of refresh control

Added the option to pass in refresh-control on the scroll-view.

Pass refreshing as a function

Pass it in as a function, instead of just a placeholder

Don't use prices-loading?

As per discussion with hesterbruikman, the spinner should be shown
only on top.

See:
https://github.com/status-im/status-react/issues/10242#issuecomment-677471250

Use shared refresh-control

Use it for the overall wallet view and for single wallet views

Show bottom sheet for mobile sync

When the user pull to refresh, show them bottom sheet for syncing over
a mobile network if the syncing is set to WiFi only.

Using :on-scroll-begin-drag instead

Since I could not find a clean way of setting :prices-loading? within
refresh-control's :on-refresh, I'm opting for this workaround.

The drawback is that as soon as the user pulls to refresh, the syncing
will start. Can you live with that?