[Fixes: #11197] Fix busy loop and don't resend group chat messages
There was an issue with the PN system that would cause a busy loop,
likely impacting performance.

Another thing that this commit addresses is that since we fixed retry of
sending messages, we noticed that it's to eager on group chats, so for
now it's disabled for them (essentially same behavior as 1.9), while
still working for 1-to-1.