Add `extractNativeLibs' flag to reduce apk size
Android APK sizes before the react-native upgrade were below 100mb and now they are 211MB.

by adding extractNativeLibs="true" in AndroidManifest.xml we are instructing the app to compress native libs and extract them after installing the apk on device.

With this change the Android APK size dropped to 103MB.

A side effect of this may be that the app may now take more space on the device.

ref : https://developer.android.com/guide/topics/manifest/application-element#extractNativeLibs

After upgrading Android Gradle Plugin to 4.2.0 and above we must use useLegacyPackaging in our app's build.gradle
and get rid of extractNativeLibs="true" from AndroidManifest.xml