Fix incorrect condition not caught by case
This is a perfect example of why it's so dangerous to use "case" with
non-literals, we can type almost any valid Clojure code in the condition. In the
diff we can see the constant (with value 101) was being called as a function.

We would expect the linter would check if the value is callable, but it doesn't
unfortunately. Therefore, not even "cond*" macros can protect us of this sort of
problem.