upgrade `react-native` to 0.73.4
fixes #18291

Changes worth mentioning are :
- More hacks/patches
- Force app to use `Java 17` everywhere to compile `kotlin,java`
- `gems` were upgraded after a long time
- `aapt2` was bumped to `8.1.1`
- `metro` is now at `0.80.4`
- `xcbeautify` was bumped to `1.4.0`
- `@react-native-community/clipboard` lib was replaced with `@react-native-clipboard/clipboard`
- `react-native-dialogs` lib was upgraded to `1.1.2`
- `react-native-gesture-handler` lib was upgraded to `2.14.1`
- `react-native-navigation` was upgraded to `7.37.2`

- Android
- iOS

every other iOS build locally with `make run-ios` would fail with this error message :

```
[React-RCTFabric] Compiling RCTThirdPartyFabricComponentsProvider.mm
❌ error: Build input file cannot be found: '/Users/siddarthkumar/code/experiments/status-mobile/
node_modules/react-native/React/Fabric/RCTThirdPartyFabricComponentsProvider.mm'.
Did you forget to declare this file as an output of a script phase or custom build rule which
produces it? (in target 'React-RCTFabric' from project 'Pods')
[React-RCTFabric] Compiling RCTTextInputComponentView.mm
⚠️ Run script build phase 'Bundle React Native code and images' will be run during every build
because it does not specify any outputs. To address this warning, either add output dependencies
to the script phase, or configure it to run in every build by unchecking "Based on dependency analysis"
 in the script phase. (in target 'StatusIm' from project 'StatusIm')
```

This gets fixed with :
- `make xcode-clean`
- `make clean`

However it is frustrating how often this happens now after upgrading to react-native 0.73 🥲