[iOS] enable new architecture for react-native
This commit enables `Fabric` and `Hermes` for `iOS` and in my initial tests the app seemed to behave well.
We can't do this for Android yet because the blur view implementation is not ported to `Fabric` properly.
needed for https://github.com/status-im/status-mobile/issues/18138

Please check builds from this PR to see if you notice any performance improvements or degradation.

Please smoke test the iOS builds to see if there are any crashes.

- iOS

status: ready