swap mp4 assets with fake ones on release APK
I recently discovered that we have `mp4` video assets as part of `Android` bundle which are never used on `Android`.
The animations are hidden only shown on certain `iOS` devices and this is done via a runtime check which includes only certain `iOS` devices.
Because of this logic and the way `react-native` bundles assets into binaries we get these assets inside the Android `APK` as well.

This commit swaps the `mp4` files with fake ones of `4 bytes` each after a release build is generated in a nix phase.
This step would only happen for `nightly`, `PR` and `release` builds, since we care about final `APK` size.

This helps us to shave off roughly `30 MB` on the `APK` size

not needed since we just optimise the size of final `APK`.

- Android

fixes: #19987

status: ready