Experiment with more solutions for using stable handlers
Re-frame docs/on-stable-dom-handlers.md offers a clever solution to cache any
function even if it needs to "close over" dynamic values.

In this commit there are two solutions, one that's independent of React, i.e. a
pure Clojure solution that should be used only with form-2 components and a
second one, that uses hooks and persists state using React refs.

Both solutions are capable of solving the problem of stable handlers. The
solution using a form-2 component is committed here as an implementation
reference since it was taken from re-frame docs. The hooks solution should be
battle tested in more components to make sure it's good enough.