chore: remove reagent override because it is currently not effective
This change removes an override for reagent's polyfill for requestAnimationFrame.
Initially, the codebase used an override to configure to use a different polyfill,
but now the override is not currently effective because of how the override is set.
The `fake-raf` def is used by the `next-tick` def inside reagent, but both of those
defs are evaluated before the override is set, meaning that the override does not take
effect since we're not re-evaluating the next-tick def.